+++ weaver/processes/wps_package.py
@@ -104,7 +104,7 @@ from weaver.vault.utils import (
     map_vault_location,
     parse_vault_token
 )
-from weaver.wps.service import ReferenceStatusLocationStorage
+from weaver.wps.storage import ReferenceStatusLocationStorage
 from weaver.wps.utils import get_wps_output_dir, get_wps_output_url, map_wps_output_location
 from weaver.wps_restapi import swagger_definitions as sd
